requirements.txt
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
imageio==2.36.0
Jinja2==3.1.4
kiwisolver==1.4.7
lpips==0.1.4
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
packaging==24.1
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
scipy==1.14.1
setuptools==75.3.0
six==1.16.0
sympy==1.13.1
-i https://download.pytorch.org/whl/cu124
torch==2.5.0
-i https://download.pytorch.org/whl/cu124
torchvision==0.20.0
torch-fidelity==0.3.0
tqdm==4.66.6
typing_extensions==4.12.2
scikit-image==0.24.0
